Extract SplitPointsTestHelper
